Rename flag cache-dir into gopath
The cache-dir flag was used to specify a cache dir for source and deps.
The fyne-cross suffix that was always added to the cache-dir, this was not allowing to use the GOPATH forcing to perform at least one dependencies download.
This commit ensure that the fyne-cross is added only for the default value allowing to specify any custom value. Additionally renames the flag name into gopath for better naming.

Fixes #4