Create dist package by default
Fyne apps are desktop ones so could make sense to have packaging
with default settings enabled by default to make the testing phase easier.

This commit removes the --dist flag and use the fyne logo as default icon if no one is specificied and the Icon.png is not present.