all: add engine flag and regression fixes
This commit fixes some regression introduced with 1.2.0 related to tags and env variable escaping.
Additionally introduces the `--engine` flags that allow to specify the container engine to use between docker and podman. The default behavior is not changed, if the flag is not specified fyne-cross will auto detect the engine.

Fixes: #96, #97